feat(web): convert billing onboarding page to modal with polished Pro upgrade wizard
Replace the /billing/onboarding route with a modal overlay on the billing page,
triggered by session_id query param. Redesign the wizard as a multi-step modal:
Welcome → Setup (unified machine+storage resize) → Domain → Complete.

- Convert BillingOnboardingPage to BillingOnboardingModal (open/onClose props)
- Merge storage and machine-size steps into one using assistantsResizeMutation
- Add Welcome and Complete celebration screens with serif typography and glow effects
- Add step progress dots in the modal footer, back navigation between steps
- Add retry button on timeout state, remove Stripe mentions from copy
- Split monolithic file into pro-onboarding/ directory (10 focused files)
- Remove /billing/onboarding route from routes.tsx
- Add CSS keyframes for onboarding animations
