+++ apps/web/src/domains/settings/billing/pro-onboarding/index.ts
@@ -1,0 +1,3 @@
+export { BillingOnboardingModal } from "./billing-onboarding-modal.js";
+export type { BillingOnboardingModalProps } from "./billing-onboarding-modal.js";
+export { allowedMachineSizesForTier, extractOnboardingErrorMessage, ONBOARDING_ERROR_CODE_MESSAGES } from "./utils.js";
